still going , need to get peers using http request and udp packet

+++ movie_streaming/bittorentProtocol/torrent.go
@@ -22,7 +22,7 @@ type FileInfo struct {
 
 type InfoMultipleFiles struct {
 	Files       []FileInfo `bencode:"files"`
-	PieceLength int        `bencode:"piece length"`
+	PieceLength int64        `bencode:"piece length"`
 	Pieces      string     `bencode:"pieces"`
 	Name        string     `bencode:"name"`
 }
